starter/app-sec-config/key-vault/python/client/keyvault_functions.py
"""
Key Vault secret management functions for storing, retrieving, and caching secrets.
These functions serve as the interface between the Flask app and Azure Key Vault.
"""
import os
import time
from azure.identity import DefaultAzureCredential
from azure.keyvault.secrets import SecretClient
from azure.core.exceptions import ResourceNotFoundError, HttpResponseError


def get_client():
    """Get a Key Vault SecretClient using Entra ID authentication."""
    vault_url = os.environ.get("KEY_VAULT_URL")

    if not vault_url:
        raise ValueError(
            "KEY_VAULT_URL environment variable must be set"
        )

    credential = DefaultAzureCredential()
    return SecretClient(vault_url=vault_url, credential=credential)


# BEGIN RETRIEVE SECRETS FUNCTION



# END RETRIEVE SECRETS FUNCTION


# BEGIN LIST SECRETS FUNCTION



# END LIST SECRETS FUNCTION


# BEGIN CREATE SECRET VERSION FUNCTION



# END CREATE SECRET VERSION FUNCTION


# BEGIN CACHED RETRIEVAL FUNCTION



# END CACHED RETRIEVAL FUNCTION
