starter/app-sec-config/key-vault/python/client/requirements.txt
azure-identity==1.21.0
azure-keyvault-secrets==4.9.0
flask==3.1.1
